Cleanup for 3.1 linux server
git-svn-id: https://unknownworlds.svn.cloudforge.com/ns1@93 67975925-1194-0748-b3d5-c16f83f1a3a1

+++ main/source/mod/AvHBaseBuildable.cpp
@@ -235,7 +235,7 @@ void AvHBaseBuildable::ConstructUse( CBaseEntity *pActivator, CBaseEntity *pCall
 				    const float kDefaultInterval = .1f;
 				    float theTimeOfLastConstructUse = thePlayer->GetTimeOfLastConstructUse();
 				    
-				    float theInterval = min(max(gpGlobals->time - theTimeOfLastConstructUse, 0), kDefaultInterval);
+				    float theInterval = min(max(gpGlobals->time - theTimeOfLastConstructUse, 0.0f), kDefaultInterval);
 				    thePercentage += (theInterval/(float)theBuildTime);
 
 				    thePlayer->SetTimeOfLastConstructUse(gpGlobals->time);
@@ -474,7 +474,7 @@ void AvHBaseBuildable::SetNormalizedBuildPercentage(float inPercentage, bool inF
 	if(theDiff > 0)
 	{
 		this->pev->health += theDiff*(1.0f - kBaseHealthPercentage)*this->mBaseHealth;
-		this->pev->health = min(max(0.0f, this->pev->health), this->mBaseHealth);
+		this->pev->health = min(max(0.0f, this->pev->health), (float)this->mBaseHealth);
 	}
 	else
 	{
